+++ examples/async_http_array_read.rs
@@ -34,7 +34,7 @@ async fn http_array_read() -> Result<(), Box<dyn std::error::Error>> {
     );
 
     // Read the whole array
-    let subset_all = ArraySubset::new_with_start_shape(vec![0, 0], array.shape().to_vec())?;
+    let subset_all = ArraySubset::new_with_shape(array.shape().to_vec());
     let data_all = array
         .async_retrieve_array_subset_ndarray::<f32>(&subset_all)
         .await?;
@@ -48,7 +48,7 @@ async fn http_array_read() -> Result<(), Box<dyn std::error::Error>> {
     println!("Chunk [1,0] is:\n{data_chunk:?}\n");
 
     // Read the central 4x2 subset of the array
-    let subset_4x2 = ArraySubset::new_with_start_shape(vec![2, 3], vec![4, 2])?; // the center 4x2 region
+    let subset_4x2 = ArraySubset::new_with_ranges(&[2..6, 3..5]); // the center 4x2 region
     let data_4x2 = array
         .async_retrieve_array_subset_ndarray::<f32>(&subset_4x2)
         .await?;