+++ src/storage/store/store_sync/http_store.rs
@@ -248,7 +248,7 @@ mod tests {
         assert_eq!(array.data_type(), &DataType::Float32);
 
         // Read the central 4x2 subset of the array
-        let subset_4x2 = ArraySubset::new_with_start_shape(vec![2, 3], vec![4, 2]).unwrap(); // the center 4x2 region
+        let subset_4x2 = ArraySubset::new_with_ranges(&[2..6, 3..5]); // the center 4x2 region
         let data_4x2 = array
             .retrieve_array_subset_elements::<f32>(&subset_4x2)
             .unwrap();
@@ -278,7 +278,7 @@ mod tests {
         assert_eq!(array.data_type(), &DataType::UInt16);
 
         // Read the central 4x2 subset of the array
-        let subset_4x2 = ArraySubset::new_with_start_shape(vec![2, 3], vec![4, 2]).unwrap(); // the center 4x2 region
+        let subset_4x2 = ArraySubset::new_with_ranges(&[2..6, 3..5]); // the center 4x2 region
         let data_4x2 = array
             .retrieve_array_subset_elements::<u16>(&subset_4x2)
             .unwrap();